# [WIP] Optimise term docs iterator