--- a/src/search/index/reader.rs
+++ b/src/search/index/reader.rs
@@ -7,7 +7,7 @@ pub trait IndexReader<'a> {
 
     fn num_docs(&self) -> usize;
     fn iter_docids_all(&'a self) -> Self::AllDocRefIterator;
-    fn iter_docids_with_term(&'a self, term: &Term, field_name: &str) -> Self::TermDocRefIterator;
+    fn iter_docids_with_term(&'a self, term: &Term, field_name: &str) -> Option<Self::TermDocRefIterator>;
     fn iter_terms(&'a self) -> Box<Iterator<Item=&'a Term> + 'a>;
     //pub fn retrieve_document(&self, &Self::DocRef) -> Document;
 }