Was previously calling a slow next_doc method on each iteration. Now uses an iterator which gives massive performance improvements (500-770x speed increase!).

Before:

```
test search::query_set::benches::bench_all                  ... bench:     113,670 ns/iter (+/- 1,169)
test search::query_set::benches::bench_buzz_term            ... bench:  12,460,039 ns/iter (+/- 96,759)
test search::query_set::benches::bench_fizz_term            ... bench:  34,493,952 ns/iter (+/- 189,039)
test search::query_set::benches::bench_fizzbuzz_conjunction ... bench:  46,930,663 ns/iter (+/- 82,442)
test search::query_set::benches::bench_fizzbuzz_disjunction ... bench:  46,987,618 ns/iter (+/- 172,257)
test search::query_set::benches::bench_fizzbuzz_exclusion   ... bench:  46,974,383 ns/iter (+/- 173,365)
```

After:
```
test search::query_set::benches::bench_all                  ... bench:     118,289 ns/iter (+/- 5,338)
test search::query_set::benches::bench_buzz_term            ... bench:      26,974 ns/iter (+/- 1,172)
test search::query_set::benches::bench_fizz_term            ... bench:      44,738 ns/iter (+/- 893)
test search::query_set::benches::bench_fizzbuzz_conjunction ... bench:      89,786 ns/iter (+/- 3,017)
test search::query_set::benches::bench_fizzbuzz_disjunction ... bench:     124,635 ns/iter (+/- 16,525)
test search::query_set::benches::bench_fizzbuzz_exclusion   ... bench:     107,404 ns/iter (+/- 5,048)
```

TODO:
 - [ ] Some error management code needs cleaning up